refactor(#7): presentation 모듈 ui 등 기타 수정
