+++ app/src/main/java/com/example/errorhandlingex/handler/TodayFortuneExceptionHandler.kt
@@ -5,7 +5,7 @@ import android.app.Application
 import android.content.Intent
 import android.os.Bundle
 import android.os.Process
-import com.example.presentation.error_handling.ErrorHandlingActivity
+import com.example.presentation.error.ErrorHandlingActivity
 import java.io.PrintWriter
 import java.io.StringWriter
 import kotlin.system.exitProcess
