+++ presentation/src/main/java/com/example/presentation/error/ErrorHandlingActivity.kt
@@ -1,4 +1,4 @@
-package com.example.presentation.error_handling
+package com.example.presentation.error
 
 import android.content.Intent
 import android.os.Bundle
