+++ presentation/src/main/res/layout/activity_main_home.xml
@@ -52,6 +52,11 @@
             android:textColor="@color/GS_700"
             android:visibility="gone" />
 
+        <ProgressBar
+            android:id="@+id/progress_bar"
+            android:layout_width="40dp"
+            android:layout_height="40dp"/>
+
         <Button
             android:id="@+id/btn_save"
             style="@style/B3_2"
